+++ lab3/helper.py
@@ -4,22 +4,8 @@
 
 __author__ = "Anton Anderzén, Stella Katsarou, Bas Straathof"
 
-import time
 import numpy as np
 import matplotlib.pyplot as plt
-
-
-def get_pict_data(fname):
-    """Gets us the data
-
-    Args:
-        fname (str): Path to file
-
-    Returns:
-        images (np.ndarray): 11 images of 32x32 pixels
-    """
-    with open(fname, 'r') as f:
-        return np.asarray(f.read().split(',')).reshape((11, 1024)).astype(int)
 
 
 def plot_images(images, ixs=None):
